requirements.txt
blinker==1.5
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pyIsEmail==2.0.1
python-dotenv==0.21.1
Werkzeug==2.2.3
